Research_TON_NFT/requirements.txt
aiogram==2.24
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
emoji==2.2.0
fake-headers==1.0.2
frozenlist==1.3.3
html5lib==1.1
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
psycopg2-binary==2.9.5
pytz==2022.7.1
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.14
webencodings==0.5.1
yarl==1.8.2
